LayoutTests/css3/flexbox/resize-min-content-flexbox.html

<!DOCTYPE html>

<link href="resources/flexbox.css" rel="stylesheet">

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<style>
body, html {
  height: 100%;
  margin: 0;
}

.flexbox.column {
  min-height: min-content;
  background: green;
  height: 100%;
}

#content {
  height: 1000px;
  max-height: 100%;
}
</style>

<div id="log"></div>

<div class="flexbox column justify-content-center align-items-center" data-expected-height="100">
  <div id="content" data-expected-height="100"></div>
</div>

<script>
  document.body.offsetHeight;
  document.documentElement.style.height = '100px';

  checkLayout('.flexbox');
</script>
